requirements/base.sav.txt
Django==1.9.6
MySQL-python==1.2.5
Pillow==3.2.0
PyJWT==1.4.0
PyYAML==3.11
apache-libcloud==1.1.0
argparse==1.2.1
bioblend==0.7.0
biopython==1.67
boto==2.39.0
colorama==0.3.7
defusedxml==0.4.1
django-admin-bootstrapped==2.5.7
django-admin-sortable==2.0.15
django-authtools==1.5.0
django-bootstrap-themes==3.3.6
django-braces==1.8.1
django-cors-headers==1.1.0
django-countries==3.4.1
django-crispy-forms==1.6.0
django-crontab==0.7.1
-e git://github.com/mvpdev/django-eav.git@f46e3433e743af3e43c28ca3abd84509e7e0b14f#egg=django_eav-master
django-environ==0.4.0
django-grappelli==2.8.1
django-jquery==1.12.2
django-jquery-ui==1.11.4.1
django-mail-templated==2.6.2
django-mptt==0.8.4
django-multiupload==0.5.1
django-nested-admin==3.0.2
django-registration==2.0.4
django-tabbed-admin==0.0.5
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
djangorestframework-xml==1.3.0
drfdocs==0.0.9
drmaa==0.7.6
easy-thumbnails==2.3
environ==1.0
inflection==0.3.1
python-monkey-business==1.0.0
pytz==2016.4
radical.utils==0.41.1
requests==2.9.1
requests-toolbelt==0.6.0
saga-python==0.41.3
six==1.10.0
sqlparse==0.1.19
wsgiref==0.1.2
